[Decomp][SpellMgr] Cohesion split: extract loaders to SpellMgrPet/Proc/Chains/Data/Areas.cpp
Pure file-cohesion split of SpellMgr.cpp (6179 -> 3965 lines). Moves 17
SpellMgr loader definitions plus their file-local helper structs into five
new SpellMgrXxx.cpp TUs; SpellMgr.h is unchanged. No behaviour change.

- SpellMgrPet.cpp:    LoadSpellPetAuras, LoadPetLevelupSpellMap,
                      LoadPetDefaultSpells_helper, LoadPetDefaultSpells
- SpellMgrProc.cpp:   LoadSpellProcEvents, LoadSpellProcItemEnchant
                      (+ SpellRankHelper template, DoSpellProcEvent,
                       DoSpellProcItemEnchant)
- SpellMgrChains.cpp: LoadSpellChains, LoadSpellLearnSkills,
                      LoadSpellLearnSpells (+ AbilitySpellPrevMap typedef,
                      LoadSpellChains_AbilityHelper)
- SpellMgrData.cpp:   LoadSpellTargetPositions, LoadSpellBonuses,
                      LoadSpellElixirs, LoadSpellThreats,
                      LoadSpellScriptTarget (+ DoSpellBonuses, DoSpellThreat,
                      SpellRankHelper template)
- SpellMgrAreas.cpp:  LoadSpellAreas, LoadSkillLineAbilityMap,
                      LoadSkillRaceClassInfoMap

CMake file(GLOB Object/*.cpp) auto-picks the new files; no CMakeLists change.
