+++ src/game/Object/SpellMgr.cpp
@@ -1538,687 +1538,6 @@ SpellCastResult GetErrorAtShapeshiftedCast(SpellEntry const* spellInfo, uint32 f
 }
 
 /**
- * @brief Loads spell target destination coordinates from the database.
- */
-void SpellMgr::LoadSpellTargetPositions()
-{
-    mSpellTargetPositions.clear();                          // need for reload case
-
-    uint32 count = 0;
-
-    //                                                0   1           2                  3                  4                  5
-    QueryResult* result = WorldDatabase.Query("SELECT `id`, `target_map`, `target_position_x`, `target_position_y`, `target_position_z`, `target_orientation` FROM `spell_target_position`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u spell target destination coordinates", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 Spell_ID = fields[0].GetUInt32();
-
-        SpellTargetPosition st;
-
-        st.target_mapId       = fields[1].GetUInt32();
-        st.target_X           = fields[2].GetFloat();
-        st.target_Y           = fields[3].GetFloat();
-        st.target_Z           = fields[4].GetFloat();
-        st.target_Orientation = fields[5].GetFloat();
-
-        MapEntry const* mapEntry = sMapStore.LookupEntry(st.target_mapId);
-        if (!mapEntry)
-        {
-            sLog.outErrorDb("Spell (ID:%u) target map (ID: %u) does not exist in `Map.dbc`.", Spell_ID, st.target_mapId);
-            continue;
-        }
-
-        if (st.target_X == 0 && st.target_Y == 0 && st.target_Z == 0)
-        {
-            sLog.outErrorDb("Spell (ID:%u) target coordinates not provided.", Spell_ID);
-            continue;
-        }
-
-        SpellEntry const* spellInfo = sSpellStore.LookupEntry(Spell_ID);
-        if (!spellInfo)
-        {
-            sLog.outErrorDb("Spell (ID:%u) listed in `spell_target_position` does not exist.", Spell_ID);
-            continue;
-        }
-
-        bool found = false;
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            SpellEffectEntry const* spellEffect = spellInfo->GetSpellEffect(SpellEffectIndex(i));
-            if (!spellEffect)
-            {
-                continue;
-            }
-            if (spellEffect->EffectImplicitTargetA==TARGET_TABLE_X_Y_Z_COORDINATES || spellEffect->EffectImplicitTargetB==TARGET_TABLE_X_Y_Z_COORDINATES)
-            {
-                found = true;
-                break;
-            }
-        }
-        if (!found)
-        {
-            sLog.outErrorDb("Spell (Id: %u) listed in `spell_target_position` does not have target TARGET_TABLE_X_Y_Z_COORDINATES (17).", Spell_ID);
-            continue;
-        }
-
-        mSpellTargetPositions[Spell_ID] = st;
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u spell target destination coordinates", count);
-    sLog.outString();
-}
-
-template <typename EntryType, typename WorkerType, typename StorageType>
-struct SpellRankHelper
-{
-    SpellRankHelper(SpellMgr& _mgr, StorageType& _storage): mgr(_mgr), worker(_storage), customRank(0) {}
-    void RecordRank(EntryType& entry, uint32 spell_id)
-    {
-        const SpellEntry* spell = sSpellStore.LookupEntry(spell_id);
-        if (!spell)
-        {
-            sLog.outErrorDb("Spell %u listed in `%s` does not exist", spell_id, worker.TableName());
-            return;
-        }
-
-        uint32 first_id = mgr.GetFirstSpellInChain(spell_id);
-
-        // most spell ranks expected same data
-        if (first_id)
-        {
-            firstRankSpells.insert(first_id);
-
-            if (first_id != spell_id)
-            {
-                if (!worker.IsValidCustomRank(entry, spell_id, first_id))
-                {
-                    return;
-                }
-                // for later check that first rank also added
-                else
-                {
-                    firstRankSpellsWithCustomRanks.insert(first_id);
-                    ++customRank;
-                }
-            }
-        }
-
-        worker.AddEntry(entry, spell);
-    }
-    void FillHigherRanks()
-    {
-        // check that first rank added for custom ranks
-        for (std::set<uint32>::const_iterator itr = firstRankSpellsWithCustomRanks.begin(); itr != firstRankSpellsWithCustomRanks.end(); ++itr)
-            if (!worker.HasEntry(*itr))
-            {
-                sLog.outErrorDb("Spell %u must be listed in `%s` as first rank for listed custom ranks of spell but not found!", *itr, worker.TableName());
-            }
-
-        // fill absent non first ranks data base at first rank data
-        for (std::set<uint32>::const_iterator itr = firstRankSpells.begin(); itr != firstRankSpells.end(); ++itr)
-        {
-            if (worker.SetStateToEntry(*itr))
-            {
-                mgr.doForHighRanks(*itr, worker);
-            }
-        }
-    }
-    std::set<uint32> firstRankSpells;
-    std::set<uint32> firstRankSpellsWithCustomRanks;
-
-    SpellMgr& mgr;
-    WorkerType worker;
-    uint32 customRank;
-};
-
-struct DoSpellProcEvent
-{
-    DoSpellProcEvent(SpellProcEventMap& _spe_map) : spe_map(_spe_map), customProc(0), count(0) {}
-    void operator()(uint32 spell_id)
-    {
-        SpellProcEventEntry const& spe = state->second;
-        // add ranks only for not filled data (some ranks have ppm data different for ranks for example)
-        SpellProcEventMap::const_iterator spellItr = spe_map.find(spell_id);
-        if (spellItr == spe_map.end())
-        {
-            spe_map[spell_id] = spe;
-        }
-        // if custom rank data added then it must be same except ppm
-        else
-        {
-            SpellProcEventEntry const& r_spe = spellItr->second;
-            if (spe.schoolMask != r_spe.schoolMask)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` as custom rank have different schoolMask from first rank in chain", spell_id);
-            }
-
-            if (spe.spellFamilyName != r_spe.spellFamilyName)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` as custom rank have different spellFamilyName from first rank in chain", spell_id);
-            }
-
-            for (int32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                if (spe.spellFamilyMask[i] != r_spe.spellFamilyMask[i])
-                {
-                    sLog.outErrorDb("Spell %u listed in `spell_proc_event` as custom rank have different spellFamilyMask/spellFamilyMask2 from first rank in chain", spell_id);
-                    break;
-                }
-            }
-
-            if (spe.procFlags != r_spe.procFlags)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` as custom rank have different procFlags from first rank in chain", spell_id);
-            }
-
-            if (spe.procEx != r_spe.procEx)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` as custom rank have different procEx from first rank in chain", spell_id);
-            }
-
-            // only ppm allowed has been different from first rank
-
-            if (spe.customChance != r_spe.customChance)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` as custom rank have different customChance from first rank in chain", spell_id);
-            }
-
-            if (spe.cooldown != r_spe.cooldown)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` as custom rank have different cooldown from first rank in chain", spell_id);
-            }
-        }
-    }
-
-    const char* TableName() { return "spell_proc_event"; }
-    bool IsValidCustomRank(SpellProcEventEntry const& spe, uint32 entry, uint32 first_id)
-    {
-        // let have independent data in table for spells with ppm rates (exist rank dependent ppm rate spells)
-        if (!spe.ppmRate)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_proc_event` is not first rank (%u) in chain", entry, first_id);
-            // prevent loading since it won't have an effect anyway
-            return false;
-        }
-        return true;
-    }
-    void AddEntry(SpellProcEventEntry const& spe, SpellEntry const* spell)
-    {
-        spe_map[spell->Id] = spe;
-
-        bool isCustom = false;
-
-        if (spe.procFlags == 0)
-        {
-            if (spell->GetProcFlags()==0)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` probally not triggered spell (no proc flags)", spell->Id);
-            }
-        }
-        else
-        {
-            if (spell->GetProcFlags()==spe.procFlags)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` has exactly same proc flags as in spell.dbc, field value redundant", spell->Id);
-            }
-            else
-            {
-                isCustom = true;
-            }
-        }
-
-        if (spe.customChance == 0)
-        {
-            /* enable for re-check cases, 0 chance ok for some cases because in some cases it set by another spell/talent spellmod)
-            if (spell->procChance==0 && !spe.ppmRate)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` probally not triggered spell (no chance or ppm)", spell->Id);
-            }
-            */
-        }
-        else
-        {
-            if (spell->GetProcChance()==spe.customChance)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` has exactly same custom chance as in spell.dbc, field value redundant", spell->Id);
-            }
-            else
-            {
-                isCustom = true;
-            }
-        }
-
-        // totally redundant record
-        if (!spe.schoolMask && !spe.procFlags &&
-            !spe.procEx && !spe.ppmRate && !spe.customChance && !spe.cooldown)
-        {
-            bool empty = !spe.spellFamilyName ? true : false;
-            for (int32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                if (spe.spellFamilyMask[i])
-                {
-                    empty = false;
-                    ClassFamilyMask const& mask = spell->GetEffectSpellClassMask(SpellEffectIndex(i));
-                    if (mask == spe.spellFamilyMask[i])
-                    {
-                        sLog.outErrorDb("Spell %u listed in `spell_proc_event` has same class mask as in Spell.dbc (EffectIndex %u) and doesn't have any other data", spell->Id, i);
-                    }
-                }
-            }
-            if (empty)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_proc_event` doesn't have any useful data", spell->Id);
-            }
-        }
-
-        if (isCustom)
-        {
-            ++customProc;
-        }
-        else
-        {
-            ++count;
-        }
-    }
-
-    bool HasEntry(uint32 spellId) { return spe_map.find(spellId) != spe_map.end(); }
-    bool SetStateToEntry(uint32 spellId) { return (state = spe_map.find(spellId)) != spe_map.end(); }
-    SpellProcEventMap& spe_map;
-    SpellProcEventMap::const_iterator state;
-
-    uint32 customProc;
-    uint32 count;
-};
-
-/**
- * @brief Loads spell proc event condition overrides from the database.
- */
-void SpellMgr::LoadSpellProcEvents()
-{
-    mSpellProcEventMap.clear();                             // need for reload case
-
-    //                                                 0        1             2                  3                    4                    5                    6                    7                    8                    9                    10                   11                   12           13        14         15              16
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`, `SchoolMask`, `SpellFamilyName`, `SpellFamilyMaskA0`, `SpellFamilyMaskA1`, `SpellFamilyMaskA2`, `SpellFamilyMaskB0`, `SpellFamilyMaskB1`, `SpellFamilyMaskB2`, `SpellFamilyMaskC0`, `SpellFamilyMaskC1`, `SpellFamilyMaskC2`, `procFlags`, `procEx`, `ppmRate`, `CustomChance`, `Cooldown` FROM `spell_proc_event`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString();
-        sLog.outString(">> No spell proc event conditions loaded");
-        return;
-    }
-
-    SpellRankHelper<SpellProcEventEntry, DoSpellProcEvent, SpellProcEventMap> rankHelper(*this, mSpellProcEventMap);
-
-    BarGoLink bar(result->GetRowCount());
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-
-        SpellProcEventEntry spe;
-
-        spe.schoolMask      = fields[1].GetUInt32();
-        spe.spellFamilyName = fields[2].GetUInt32();
-
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            spe.spellFamilyMask[i] = ClassFamilyMask(
-                                         (uint64)fields[i + 3].GetUInt32() | ((uint64)fields[i + 6].GetUInt32() << 32),
-                                         fields[i + 9].GetUInt32());
-        }
-        spe.procFlags       = fields[12].GetUInt32();
-        spe.procEx          = fields[13].GetUInt32();
-        spe.ppmRate         = fields[14].GetFloat();
-        spe.customChance    = fields[15].GetFloat();
-        spe.cooldown        = fields[16].GetUInt32();
-
-        rankHelper.RecordRank(spe, entry);
-    }
-    while (result->NextRow());
-
-    rankHelper.FillHigherRanks();
-
-    delete result;
-
-    sLog.outString(">> Loaded %u extra spell proc event conditions +%u custom proc (inc. +%u custom ranks)",  rankHelper.worker.count, rankHelper.worker.customProc, rankHelper.customRank);
-    sLog.outString();
-}
-
-struct DoSpellProcItemEnchant
-{
-    DoSpellProcItemEnchant(SpellProcItemEnchantMap& _procMap, float _ppm) : procMap(_procMap), ppm(_ppm) {}
-    void operator()(uint32 spell_id) { procMap[spell_id] = ppm; }
-
-    SpellProcItemEnchantMap& procMap;
-    float ppm;
-};
-
-/**
- * @brief Loads proc-per-minute data for spell item enchant procs.
- */
-void SpellMgr::LoadSpellProcItemEnchant()
-{
-    mSpellProcItemEnchantMap.clear();                       // need for reload case
-
-    uint32 count = 0;
-
-    //                                                0      1
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`, `ppmRate` FROM `spell_proc_item_enchant`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u proc item enchant definitions", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-        float ppmRate = fields[1].GetFloat();
-
-        SpellEntry const* spellInfo = sSpellStore.LookupEntry(entry);
-
-        if (!spellInfo)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_proc_item_enchant` does not exist", entry);
-            continue;
-        }
-
-        uint32 first_id = GetFirstSpellInChain(entry);
-
-        if (first_id != entry)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_proc_item_enchant` is not first rank (%u) in chain", entry, first_id);
-            // prevent loading since it won't have an effect anyway
-            continue;
-        }
-
-        mSpellProcItemEnchantMap[entry] = ppmRate;
-
-        // also add to high ranks
-        DoSpellProcItemEnchant worker(mSpellProcItemEnchantMap, ppmRate);
-        doForHighRanks(entry, worker);
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u proc item enchant definitions", count);
-    sLog.outString();
-}
-
-bool IsCastEndProcModifierAura(SpellEntry const* spellInfo, SpellEffectIndex effecIdx, SpellEntry const* procSpell)
-{
-    SpellEffectEntry const* spellEffect = spellInfo->GetSpellEffect(SpellEffectIndex(effecIdx));
-    if (!spellEffect)
-    {
-        return false;
-    }
-
-    // modifier auras that can proc on cast end
-    switch (AuraType(spellEffect->EffectApplyAuraName))
-    {
-        case SPELL_AURA_ADD_FLAT_MODIFIER:
-        case SPELL_AURA_ADD_PCT_MODIFIER:
-        {
-            switch (spellEffect->EffectMiscValue)
-            {
-                case SPELLMOD_RANGE:
-                case SPELLMOD_RADIUS:
-                case SPELLMOD_CRITICAL_CHANCE:
-                case SPELLMOD_NOT_LOSE_CASTING_TIME:
-                case SPELLMOD_CASTING_TIME:
-                case SPELLMOD_COOLDOWN:
-                case SPELLMOD_COST:
-                case SPELLMOD_GLOBAL_COOLDOWN:
-                    return true;
-                default:
-                    break;
-            }
-        }
-        case SPELL_AURA_MOD_DAMAGE_PERCENT_DONE:
-        {
-            for (int32 i = 0; i < MAX_EFFECT_INDEX; ++i)
-                if (IsEffectHandledOnDelayedSpellLaunch(procSpell, SpellEffectIndex(i)))
-                {
-                    return true;
-                }
-
-            return false;
-        }
-        default:
-            return false;
-    }
-}
-
-struct DoSpellBonuses
-{
-    DoSpellBonuses(SpellBonusMap& _spellBonusMap, SpellBonusEntry const& _spellBonus) : spellBonusMap(_spellBonusMap), spellBonus(_spellBonus) {}
-    void operator()(uint32 spell_id) { spellBonusMap[spell_id] = spellBonus; }
-
-    SpellBonusMap& spellBonusMap;
-    SpellBonusEntry const& spellBonus;
-};
-
-/**
- * @brief Loads spell bonus coefficient overrides from the database.
- */
-void SpellMgr::LoadSpellBonuses()
-{
-    mSpellBonusMap.clear();                             // need for reload case
-    uint32 count = 0;
-    QueryResult* result = WorldDatabase.Query("SELECT entry, direct_bonus, \
-        `dot_bonus`, `ap_bonus`, `ap_dot_bonus` FROM `spell_bonus_data`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u spell bonus data", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-        uint32 entry = fields[0].GetUInt32();
-
-        SpellEntry const* spell = sSpellStore.LookupEntry(entry);
-        if (!spell)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_bonus_data` does not exist", entry);
-            continue;
-        }
-
-        uint32 first_id = GetFirstSpellInChain(entry);
-
-        if (first_id != entry)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_bonus_data` is not first rank (%u) in chain", entry, first_id);
-            // prevent loading since it won't have an effect anyway
-            continue;
-        }
-
-        SpellBonusEntry sbe;
-
-        sbe.direct_damage = fields[1].GetFloat();
-        sbe.dot_damage    = fields[2].GetFloat();
-        sbe.ap_bonus      = fields[3].GetFloat();
-        sbe.ap_dot_bonus   = fields[4].GetFloat();
-
-        bool need_dot = false;
-        bool need_direct = false;
-        uint32 x = 0;                                       // count all, including empty, meaning: not all existing effect is DoTs/HoTs
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            SpellEffectEntry const* spellEffect = spell->GetSpellEffect(SpellEffectIndex(i));
-            if (!spellEffect)
-            {
-                continue;
-            }
-            if (!spellEffect->Effect)
-            {
-                ++x;
-                continue;
-            }
-
-            // DoTs/HoTs
-            switch(spellEffect->EffectApplyAuraName)
-            {
-                case SPELL_AURA_PERIODIC_DAMAGE:
-                case SPELL_AURA_PERIODIC_DAMAGE_PERCENT:
-                case SPELL_AURA_PERIODIC_LEECH:
-                case SPELL_AURA_PERIODIC_HEAL:
-                case SPELL_AURA_OBS_MOD_HEALTH:
-                case SPELL_AURA_PERIODIC_MANA_LEECH:
-                case SPELL_AURA_OBS_MOD_MANA:
-                case SPELL_AURA_POWER_BURN_MANA:
-                    need_dot = true;
-                    ++x;
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        // TODO: maybe add explicit list possible direct damage spell effects...
-        if (x < MAX_EFFECT_INDEX)
-        {
-            need_direct = true;
-        }
-
-        // Check if direct_bonus is needed in `spell_bonus_data`
-        float direct_calc = 0.0f;
-        float direct_diff = 1000.0f;                        // for have big diff if no DB field value
-        if (sbe.direct_damage)
-        {
-            bool isHeal = false;
-            for(int i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                SpellEffectEntry const* spellEffect = spell->GetSpellEffect(SpellEffectIndex(i));
-                if (!spellEffect)
-                {
-                    continue;
-                }
-                // Heals (Also count Mana Shield and Absorb effects as heals)
-                if (spellEffect->Effect == SPELL_EFFECT_HEAL || spellEffect->Effect == SPELL_EFFECT_HEAL_MAX_HEALTH ||
-                    (spellEffect->Effect == SPELL_EFFECT_APPLY_AURA && (spellEffect->EffectApplyAuraName == SPELL_AURA_SCHOOL_ABSORB || spellEffect->EffectApplyAuraName == SPELL_AURA_PERIODIC_HEAL)) )
-                {
-                    isHeal = true;
-                    break;
-                }
-            }
-            direct_calc = CalculateDefaultCoefficient(spell, SPELL_DIRECT_DAMAGE) * (isHeal ? SCALE_SPELLPOWER_HEALING : 1.0f);
-            direct_diff = std::abs(sbe.direct_damage - direct_calc);
-        }
-
-        // Check if dot_bonus is needed in `spell_bonus_data`
-        float dot_calc = 0.0f;
-        float dot_diff = 1000.0f;                           // for have big diff if no DB field value
-        if (sbe.dot_damage)
-        {
-            bool isHeal = false;
-            for(int i = 0; i < MAX_EFFECT_INDEX; ++i)
-            {
-                SpellEffectEntry const* spellEffect = spell->GetSpellEffect(SpellEffectIndex(i));
-                if (!spellEffect)
-                {
-                    continue;
-                }
-                // Periodic Heals
-                if (spellEffect->Effect == SPELL_EFFECT_APPLY_AURA && spellEffect->EffectApplyAuraName == SPELL_AURA_PERIODIC_HEAL)
-                {
-                    isHeal = true;
-                    break;
-                }
-            }
-            dot_calc = CalculateDefaultCoefficient(spell, DOT) * (isHeal ? SCALE_SPELLPOWER_HEALING : 1.0f);
-            dot_diff = std::abs(sbe.dot_damage - dot_calc);
-        }
-
-        // direct bonus
-        if (direct_diff < 0.02f && !need_dot && !sbe.ap_bonus && !sbe.ap_dot_bonus)
-            sLog.outErrorDb("`spell_bonus_data` entry for spell %u `direct_bonus` not needed (data from table: %f, calculated %f, difference of %f) and `dot_bonus` also not used",
-                            entry, sbe.direct_damage, direct_calc, direct_diff);
-        else if (direct_diff < 0.02f && dot_diff < 0.02f && !sbe.ap_bonus && !sbe.ap_dot_bonus)
-        {
-            sLog.outErrorDb("`spell_bonus_data` entry for spell %u `direct_bonus` not needed (data from table: %f, calculated %f, difference of %f) and ",
-                            entry, sbe.direct_damage, direct_calc, direct_diff);
-            sLog.outErrorDb("                                  ... `dot_bonus` not needed (data from table: %f, calculated %f, difference of %f)",
-                            sbe.dot_damage, dot_calc, dot_diff);
-        }
-        else if (!need_direct && dot_diff < 0.02f && !sbe.ap_bonus && !sbe.ap_dot_bonus)
-            sLog.outErrorDb("`spell_bonus_data` entry for spell %u `dot_bonus` not needed (data from table: %f, calculated %f, difference of %f) and direct also not used",
-                            entry, sbe.dot_damage, dot_calc, dot_diff);
-        else if (!need_direct && sbe.direct_damage)
-        {
-            sLog.outErrorDb("`spell_bonus_data` entry for spell %u `direct_bonus` not used (spell not have non-periodic affects)", entry);
-        }
-        else if (!need_dot && sbe.dot_damage)
-        {
-            sLog.outErrorDb("`spell_bonus_data` entry for spell %u `dot_bonus` not used (spell not have periodic affects)", entry);
-        }
-
-        if (!need_direct && sbe.ap_bonus)
-        {
-            sLog.outErrorDb("`spell_bonus_data` entry for spell %u `ap_bonus` not used (spell not have non-periodic affects)", entry);
-        }
-        else if (!need_dot && sbe.ap_dot_bonus)
-        {
-            sLog.outErrorDb("`spell_bonus_data` entry for spell %u `ap_dot_bonus` not used (spell not have periodic affects)", entry);
-        }
-
-        mSpellBonusMap[entry] = sbe;
-
-        // also add to high ranks
-        DoSpellBonuses worker(mSpellBonusMap, sbe);
-        doForHighRanks(entry, worker);
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u extra spell bonus data",  count);
-    sLog.outString();
-}
-
-/**
  * @brief Checks whether a proc event definition can be triggered by a proc context.
  *
  * @param spellProcEvent The proc event definition to evaluate.
@@ -2306,165 +1625,6 @@ bool SpellMgr::IsSpellProcEventCanTriggeredBy(SpellProcEventEntry const* spellPr
         }
     }
     return false;
-}
-
-/**
- * @brief Loads elixir classification data from the database.
- */
-void SpellMgr::LoadSpellElixirs()
-{
-    mSpellElixirs.clear();                                  // need for reload case
-
-    uint32 count = 0;
-
-    //                                                0      1
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`, `mask` FROM `spell_elixir`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-
-        bar.step();
-
-        sLog.outString(">> Loaded %u spell elixir definitions", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-        uint8 mask = fields[1].GetUInt8();
-
-        SpellEntry const* spellInfo = sSpellStore.LookupEntry(entry);
-
-        if (!spellInfo)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_elixir` does not exist", entry);
-            continue;
-        }
-
-        mSpellElixirs[entry] = mask;
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u spell elixir definitions", count);
-    sLog.outString();
-}
-
-struct DoSpellThreat
-{
-    DoSpellThreat(SpellThreatMap& _threatMap) : threatMap(_threatMap), count(0) {}
-    void operator()(uint32 spell_id)
-    {
-        SpellThreatEntry const& ste = state->second;
-        // add ranks only for not filled data (spells adding flat threat are usually different for ranks)
-        SpellThreatMap::const_iterator spellItr = threatMap.find(spell_id);
-        if (spellItr == threatMap.end())
-        {
-            threatMap[spell_id] = ste;
-        }
-
-        // just assert that entry is not redundant
-        else
-        {
-            SpellThreatEntry const& r_ste = spellItr->second;
-            if (ste.threat == r_ste.threat && ste.multiplier == r_ste.multiplier && ste.ap_bonus == r_ste.ap_bonus)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_threat` as custom rank has same data as Rank 1, so redundant", spell_id);
-            }
-        }
-    }
-    const char* TableName() { return "spell_threat"; }
-    bool IsValidCustomRank(SpellThreatEntry const& ste, uint32 entry, uint32 first_id)
-    {
-        if (!ste.threat)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_threat` is not first rank (%u) in chain and has no threat", entry, first_id);
-            // prevent loading unexpected data
-            return false;
-        }
-        return true;
-    }
-    void AddEntry(SpellThreatEntry const& ste, SpellEntry const* spell)
-    {
-        threatMap[spell->Id] = ste;
-
-        // flat threat bonus and attack power bonus currently only work properly when all
-        // effects have same targets, otherwise, we'd need to seperate it by effect index
-        if (ste.threat || ste.ap_bonus != 0.f)
-        {
-            SpellEffectEntry const* spellEffect0 = spell->GetSpellEffect(EFFECT_INDEX_0);
-            SpellEffectEntry const* spellEffect1 = spell->GetSpellEffect(EFFECT_INDEX_1);
-            SpellEffectEntry const* spellEffect2 = spell->GetSpellEffect(EFFECT_INDEX_2);
-            if ((spellEffect1 && spellEffect1->EffectImplicitTargetA && (!spellEffect0 || spellEffect1->EffectImplicitTargetA != spellEffect0->EffectImplicitTargetA)) ||
-                (spellEffect2 && spellEffect2->EffectImplicitTargetA && (!spellEffect0 || spellEffect2->EffectImplicitTargetA != spellEffect0->EffectImplicitTargetA)))
-                sLog.outErrorDb("Spell %u listed in `spell_threat` has effects with different targets, threat may be assigned incorrectly", spell->Id);
-        }
-        ++count;
-    }
-    bool HasEntry(uint32 spellId) { return threatMap.count(spellId) > 0; }
-    bool SetStateToEntry(uint32 spellId) { return (state = threatMap.find(spellId)) != threatMap.end(); }
-
-    SpellThreatMap& threatMap;
-    SpellThreatMap::const_iterator state;
-    uint32 count;
-};
-
-/**
- * @brief Loads custom spell threat definitions from the database.
- */
-void SpellMgr::LoadSpellThreats()
-{
-    mSpellThreatMap.clear();                                // need for reload case
-
-    //                                                0      1       2           3
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`, `Threat`, `multiplier`, `ap_bonus` FROM `spell_threat`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> No spell threat entries loaded.");
-        sLog.outString();
-        return;
-    }
-
-    SpellRankHelper<SpellThreatEntry, DoSpellThreat, SpellThreatMap> rankHelper(*this, mSpellThreatMap);
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 entry = fields[0].GetUInt32();
-
-        SpellThreatEntry ste;
-        ste.threat = fields[1].GetUInt16();
-        ste.multiplier = fields[2].GetFloat();
-        ste.ap_bonus = fields[3].GetFloat();
-
-        rankHelper.RecordRank(ste, entry);
-    }
-    while (result->NextRow());
-
-    rankHelper.FillHigherRanks();
-
-    delete result;
-
-    sLog.outString(">> Loaded %u spell threat entries", rankHelper.worker.count);
-    sLog.outString();
 }
 
 /**
@@ -3771,1086 +2931,6 @@ SpellEntry const* SpellMgr::SelectAuraRankForLevel(SpellEntry const* spellInfo, 
     return NULL;
 }
 
-typedef std::unordered_map<uint32, uint32> AbilitySpellPrevMap;
-
-/**
- * @brief Recursively builds spell chain links from ability rank data.
- *
- * @param chainMap The spell chain map being populated.
- * @param prevRanks The previous-rank lookup table.
- * @param spell_id The current spell id.
- * @param prev_id The previous spell id in the chain.
- * @param deep The remaining recursion depth guard.
- */
-static void LoadSpellChains_AbilityHelper(SpellChainMap& chainMap, AbilitySpellPrevMap const& prevRanks, uint32 spell_id, uint32 prev_id, uint32 deep = 30)
-{
-    // spell already listed in chains store
-    SpellChainMap::const_iterator chain_itr = chainMap.find(spell_id);
-    if (chain_itr != chainMap.end())
-    {
-        MANGOS_ASSERT(chain_itr->second.prev == prev_id && "LoadSpellChains_AbilityHelper: Conflicting data in talents or spell abilities dbc");
-        return;
-    }
-
-    // prev rank listed in main chain table (can fill correct data directly)
-    SpellChainMap::const_iterator prev_chain_itr = chainMap.find(prev_id);
-    if (prev_chain_itr != chainMap.end())
-    {
-        SpellChainNode node;
-        node.prev  = prev_id;
-        node.first = prev_chain_itr->second.first;
-        node.rank  = prev_chain_itr->second.rank + 1;
-        node.req   = 0;
-        chainMap[spell_id] = node;
-        return;
-    }
-
-    // prev spell not listed in prev ranks store, so it first rank
-    AbilitySpellPrevMap::const_iterator prev_itr = prevRanks.find(prev_id);
-    if (prev_itr == prevRanks.end())
-    {
-        SpellChainNode prev_node;
-        prev_node.prev  = 0;
-        prev_node.first = prev_id;
-        prev_node.rank  = 1;
-        prev_node.req   = 0;
-        chainMap[prev_id] = prev_node;
-
-        SpellChainNode node;
-        node.prev  = prev_id;
-        node.first = prev_id;
-        node.rank  = 2;
-        node.req   = 0;
-        chainMap[spell_id] = node;
-        return;
-    }
-
-    if (deep == 0)
-    {
-        MANGOS_ASSERT(false && "LoadSpellChains_AbilityHelper: Infinity cycle in spell ability data");
-        return;
-    }
-
-    // prev rank listed, so process it first
-    LoadSpellChains_AbilityHelper(chainMap, prevRanks, prev_id, prev_itr->second, deep - 1);
-
-    // prev rank must be listed now
-    prev_chain_itr = chainMap.find(prev_id);
-    if (prev_chain_itr == chainMap.end())
-    {
-        return;
-    }
-
-    SpellChainNode node;
-    node.prev  = prev_id;
-    node.first = prev_chain_itr->second.first;
-    node.rank  = prev_chain_itr->second.rank + 1;
-    node.req   = 0;
-    chainMap[spell_id] = node;
-}
-
-/**
- * @brief Loads spell rank chain data from DBC and database sources.
- */
-void SpellMgr::LoadSpellChains()
-{
-    mSpellChains.clear();                                   // need for reload case
-    mSpellChainsNext.clear();                               // need for reload case
-
-    // load known data for talents
-    for (unsigned int i = 0; i < sTalentStore.GetNumRows(); ++i)
-    {
-        TalentEntry const* talentInfo = sTalentStore.LookupEntry(i);
-        if (!talentInfo)
-        {
-            continue;
-        }
-
-        // not add ranks for 1 ranks talents (if exist non ranks spells then it will included in table data)
-        if (!talentInfo->RankID[1])
-        {
-            continue;
-        }
-
-        for (int j = 0; j < MAX_TALENT_RANK; ++j)
-        {
-            uint32 spell_id = talentInfo->RankID[j];
-            if (!spell_id)
-            {
-                continue;
-            }
-
-            if (!sSpellStore.LookupEntry(spell_id))
-            {
-                // sLog.outErrorDb("Talent %u not exist as spell",spell_id);
-                continue;
-            }
-
-            SpellChainNode node;
-            node.prev  = (j > 0) ? talentInfo->RankID[j - 1] : 0;
-            node.first = talentInfo->RankID[0];
-            node.rank  = j + 1;
-            node.req   = 0;
-
-            mSpellChains[spell_id] = node;
-        }
-    }
-
-    // load known data from spell abilities
-    {
-        // we can calculate ranks only after full data generation
-        AbilitySpellPrevMap prevRanks;
-        for (SkillLineAbilityMap::const_iterator ab_itr = mSkillLineAbilityMap.begin(); ab_itr != mSkillLineAbilityMap.end(); ++ab_itr)
-        {
-            uint32 spell_id = ab_itr->first;
-
-            // skip GM/test/internal spells.begin Its not have ranks anyway
-            if (ab_itr->second->skillId == SKILL_INTERNAL)
-            {
-                continue;
-            }
-
-            // some forward spells not exist and can be ignored (some outdated data)
-            SpellEntry const* spell_entry = sSpellStore.LookupEntry(spell_id);
-            if (!spell_entry)                               // no cases
-            {
-                continue;
-            }
-
-            // ignore spell without forwards (non ranked or missing info in skill abilities)
-            uint32 forward_id = ab_itr->second->forward_spellid;
-            if (!forward_id)
-            {
-                continue;
-            }
-
-            // some forward spells not exist and can be ignored (some outdated data)
-            SpellEntry const* forward_entry = sSpellStore.LookupEntry(forward_id);
-            if (!forward_entry)
-            {
-                continue;
-            }
-
-            // some forward spells still exist but excluded from real use as ranks and not listed in skill abilities now
-            SkillLineAbilityMapBounds bounds = mSkillLineAbilityMap.equal_range(forward_id);
-            if (bounds.first == bounds.second)
-            {
-                continue;
-            }
-
-            // spell already listed in chains store
-            SpellChainMap::const_iterator chain_itr = mSpellChains.find(forward_id);
-            if (chain_itr != mSpellChains.end())
-            {
-                MANGOS_ASSERT(chain_itr->second.prev == spell_id && "Conflicting data in talents or spell abilities dbc");
-                continue;
-            }
-
-            // spell already listed in prev ranks store
-            AbilitySpellPrevMap::const_iterator prev_itr = prevRanks.find(forward_id);
-            if (prev_itr != prevRanks.end())
-            {
-                MANGOS_ASSERT(prev_itr->second == spell_id && "Conflicting data in talents or spell abilities dbc");
-                continue;
-            }
-
-            // prev rank listed in main chain table (can fill correct data directly)
-            SpellChainMap::const_iterator prev_chain_itr = mSpellChains.find(spell_id);
-            if (prev_chain_itr != mSpellChains.end())
-            {
-                SpellChainNode node;
-                node.prev  = spell_id;
-                node.first = prev_chain_itr->second.first;
-                node.rank  = prev_chain_itr->second.rank + 1;
-                node.req   = 0;
-
-                mSpellChains[forward_id] = node;
-                continue;
-            }
-
-            // need temporary store for later rank calculation
-            prevRanks[forward_id] = spell_id;
-        }
-
-        while (!prevRanks.empty())
-        {
-            uint32 spell_id = prevRanks.begin()->first;
-            uint32 prev_id  = prevRanks.begin()->second;
-            prevRanks.erase(prevRanks.begin());
-
-            LoadSpellChains_AbilityHelper(mSpellChains, prevRanks, spell_id, prev_id);
-        }
-    }
-
-    // load custom case
-    QueryResult* result = WorldDatabase.Query("SELECT `spell_id`, `prev_spell`, `first_spell`, `rank`, `req_spell` FROM `spell_chain`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-
-        sLog.outString(">> Loaded 0 spell chain records");
-        sLog.outErrorDb("`spell_chains` table is empty!");
-        sLog.outString();
-        return;
-    }
-
-    uint32 dbc_count = mSpellChains.size();
-    uint32 new_count = 0;
-    uint32 req_count = 0;
-
-    BarGoLink bar(result->GetRowCount());
-    do
-    {
-        bar.step();
-        Field* fields = result->Fetch();
-
-        uint32 spell_id = fields[0].GetUInt32();
-
-        SpellChainNode node;
-        node.prev  = fields[1].GetUInt32();
-        node.first = fields[2].GetUInt32();
-        node.rank  = fields[3].GetUInt8();
-        node.req   = fields[4].GetUInt32();
-
-        if (!sSpellStore.LookupEntry(spell_id))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_chain` does not exist", spell_id);
-            continue;
-        }
-
-        SpellChainMap::iterator chain_itr = mSpellChains.find(spell_id);
-        if (chain_itr != mSpellChains.end())
-        {
-            if (chain_itr->second.rank != node.rank)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` expected rank %u by DBC data.",
-                                spell_id, node.prev, node.first, node.rank, node.req, chain_itr->second.rank);
-                continue;
-            }
-
-            if (chain_itr->second.prev != node.prev)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` expected prev %u by DBC data.",
-                                spell_id, node.prev, node.first, node.rank, node.req, chain_itr->second.prev);
-                continue;
-            }
-
-            if (chain_itr->second.first != node.first)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` expected first %u by DBC data.",
-                                spell_id, node.prev, node.first, node.rank, node.req, chain_itr->second.first);
-                continue;
-            }
-
-            // update req field by table data
-            if (node.req)
-            {
-                chain_itr->second.req = node.req;
-                ++req_count;
-                continue;
-            }
-
-            // in other case redundant
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) already added (talent or spell ability with forward) and non need in `spell_chain`",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        if (node.prev != 0 && !sSpellStore.LookupEntry(node.prev))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has nonexistent previous rank spell.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        if (!sSpellStore.LookupEntry(node.first))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not existing first rank spell.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        // check basic spell chain data integrity (note: rank can be equal 0 or 1 for first/single spell)
-        if ((spell_id == node.first) != (node.rank <= 1) ||
-                (spell_id == node.first) != (node.prev == 0) ||
-                (node.rank <= 1) != (node.prev == 0))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not compatible chain data.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        if (node.req != 0 && !sSpellStore.LookupEntry(node.req))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not existing required spell.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        // talents not required data in spell chain for work, but must be checked if present for integrity
-        if (TalentSpellPos const* pos = GetTalentSpellPos(spell_id))
-        {
-            if (node.rank != pos->rank + 1)
-            {
-                sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong rank.",
-                                spell_id, node.prev, node.first, node.rank, node.req);
-                continue;
-            }
-
-            if (TalentEntry const* talentEntry = sTalentStore.LookupEntry(pos->talent_id))
-            {
-                if (node.first != talentEntry->RankID[0])
-                {
-                    sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong first rank spell.",
-                                    spell_id, node.prev, node.first, node.rank, node.req);
-                    continue;
-                }
-
-                if (node.rank > 1 && node.prev != talentEntry->RankID[node.rank - 1 - 1])
-                {
-                    sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong prev rank spell.",
-                                    spell_id, node.prev, node.first, node.rank, node.req);
-                    continue;
-                }
-
-                /*if (node.req!=talentEntry->DependsOnSpell)
-                {
-                    sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong required spell.",
-                        spell_id,node.prev,node.first,node.rank,node.req);
-                    continue;
-                }*/
-            }
-        }
-
-        // removed ranks often still listed as forward in skill abilities but not listed as spell in it
-        if (node.prev)
-        {
-            bool skip = false;
-            // some forward spells still exist but excluded from real use as ranks and not listed in skill abilities now
-            SkillLineAbilityMapBounds bounds = mSkillLineAbilityMap.equal_range(spell_id);
-            if (bounds.first == bounds.second)
-            {
-                SkillLineAbilityMapBounds prev_bounds = mSkillLineAbilityMap.equal_range(node.prev);
-                for (SkillLineAbilityMap::const_iterator ab_itr = prev_bounds.first; ab_itr != prev_bounds.second; ++ab_itr)
-                {
-                    // spell listed as forward and not listed as ability
-                    // this is marker for removed ranks
-                    if (ab_itr->second->forward_spellid == spell_id)
-                    {
-                        sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` is removed rank by DBC data.",
-                                        spell_id, node.prev, node.first, node.rank, node.req);
-                        skip = true;
-                        break;
-                    }
-                }
-            }
-
-            if (skip)
-            {
-                continue;
-            }
-        }
-
-        mSpellChains[spell_id] = node;
-
-        ++new_count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    // additional integrity checks
-    for (SpellChainMap::const_iterator i = mSpellChains.begin(); i != mSpellChains.end(); ++i)
-    {
-        if (i->second.prev)
-        {
-            SpellChainMap::const_iterator i_prev = mSpellChains.find(i->second.prev);
-            if (i_prev == mSpellChains.end())
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not found previous rank spell in table.",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req);
-            }
-            else if (i_prev->second.first != i->second.first)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has different first spell in chain compared to previous rank spell (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_prev->second.prev, i_prev->second.first, i_prev->second.rank, i_prev->second.req);
-            }
-            else if (i_prev->second.rank + 1 != i->second.rank)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has different rank compared to previous rank spell (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_prev->second.prev, i_prev->second.first, i_prev->second.rank, i_prev->second.req);
-            }
-        }
-
-        if (i->second.req)
-        {
-            SpellChainMap::const_iterator i_req = mSpellChains.find(i->second.req);
-            if (i_req == mSpellChains.end())
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not found required rank spell in table.",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req);
-            }
-            else if (i_req->second.first == i->second.first)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has required rank spell from same spell chain (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_req->second.prev, i_req->second.first, i_req->second.rank, i_req->second.req);
-            }
-            else if (i_req->second.req)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has required rank spell with required spell (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_req->second.prev, i_req->second.first, i_req->second.rank, i_req->second.req);
-            }
-        }
-    }
-
-    // fill next rank cache
-    for (SpellChainMap::const_iterator i = mSpellChains.begin(); i != mSpellChains.end(); ++i)
-    {
-        uint32 spell_id = i->first;
-        SpellChainNode const& node = i->second;
-
-        if (node.prev)
-        {
-            mSpellChainsNext.insert(SpellChainMapNext::value_type(node.prev, spell_id));
-        }
-
-        if (node.req)
-        {
-            mSpellChainsNext.insert(SpellChainMapNext::value_type(node.req, spell_id));
-        }
-    }
-
-    // check single rank redundant cases (single rank talents/spell abilities not added by default so this can be only custom cases)
-    for (SpellChainMap::const_iterator i = mSpellChains.begin(); i != mSpellChains.end(); ++i)
-    {
-        // skip non-first ranks, and spells with additional reqs
-        if (i->second.rank > 1 || i->second.req)
-        {
-            continue;
-        }
-
-        if (mSpellChainsNext.find(i->first) == mSpellChainsNext.end())
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has single rank data, so redundant.",
-                            i->first, i->second.prev, i->second.first, i->second.rank, i->second.req);
-        }
-    }
-
-    sLog.outString(">> Loaded %u spell chain records (%u from DBC data with %u req field updates, and %u loaded from table)", dbc_count + new_count, dbc_count, req_count, new_count);
-    sLog.outString();
-}
-
-/**
- * @brief Builds the spell-to-skill-learning map from DBC data.
- */
-void SpellMgr::LoadSpellLearnSkills()
-{
-    mSpellLearnSkills.clear();                              // need for reload case
-
-    // search auto-learned skills and add its to map also for use in unlearn spells/talents
-    uint32 dbc_count = 0;
-    BarGoLink bar(sSpellStore.GetNumRows());
-    for (uint32 spell = 0; spell < sSpellStore.GetNumRows(); ++spell)
-    {
-        bar.step();
-        SpellEntry const* entry = sSpellStore.LookupEntry(spell);
-
-        if (!entry)
-        {
-            continue;
-        }
-
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            SpellEffectEntry const* spellEffect = entry->GetSpellEffect(SpellEffectIndex(i));
-            if (!spellEffect)
-            {
-                continue;
-            }
-            if (spellEffect->Effect == SPELL_EFFECT_SKILL)
-            {
-                SpellLearnSkillNode dbc_node;
-                dbc_node.skill    = spellEffect->EffectMiscValue;
-                dbc_node.step     = entry->CalculateSimpleValue(SpellEffectIndex(i));
-                if (dbc_node.skill != SKILL_RIDING)
-                {
-                    dbc_node.value = 1;
-                }
-                else
-                {
-                    dbc_node.value = dbc_node.step * 75;
-                }
-                dbc_node.maxvalue = dbc_node.step * 75;
-
-                mSpellLearnSkills[spell] = dbc_node;
-                ++dbc_count;
-                break;
-            }
-        }
-    }
-
-    sLog.outString(">> Loaded %u Spell Learn Skills from DBC", dbc_count);
-    sLog.outString();
-}
-
-/**
- * @brief Loads explicit and DBC-derived learned-spell relationships.
- */
-void SpellMgr::LoadSpellLearnSpells()
-{
-    mSpellLearnSpells.clear();                              // need for reload case
-
-    //                                                0      1        2
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`, `SpellID`, `Active` FROM `spell_learn_spell`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-
-        sLog.outString();
-        sLog.outString(">> Loaded 0 spell learn spells");
-        sLog.outErrorDb("`spell_learn_spell` table is empty!");
-        return;
-    }
-
-    uint32 count = 0;
-
-    BarGoLink bar(result->GetRowCount());
-    do
-    {
-        bar.step();
-        Field* fields = result->Fetch();
-
-        uint32 spell_id    = fields[0].GetUInt32();
-
-        SpellLearnSpellNode node;
-        node.spell      = fields[1].GetUInt32();
-        node.active     = fields[2].GetBool();
-        node.autoLearned = false;
-
-        if (!sSpellStore.LookupEntry(spell_id))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_learn_spell` does not exist", spell_id);
-            continue;
-        }
-
-        if (!sSpellStore.LookupEntry(node.spell))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_learn_spell` learning nonexistent spell %u", spell_id, node.spell);
-            continue;
-        }
-
-        if (GetTalentSpellCost(node.spell))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_learn_spell` attempt learning talent spell %u, skipped", spell_id, node.spell);
-            continue;
-        }
-
-        mSpellLearnSpells.insert(SpellLearnSpellMap::value_type(spell_id, node));
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    // search auto-learned spells and add its to map also for use in unlearn spells/talents
-    uint32 dbc_count = 0;
-    for (uint32 spell = 0; spell < sSpellStore.GetNumRows(); ++spell)
-    {
-        SpellEntry const* entry = sSpellStore.LookupEntry(spell);
-
-        if (!entry)
-        {
-            continue;
-        }
-
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            SpellEffectEntry const* spellEffect = entry->GetSpellEffect(SpellEffectIndex(i));
-            if (!spellEffect)
-            {
-                continue;
-            }
-            if (spellEffect->Effect == SPELL_EFFECT_LEARN_SPELL)
-            {
-                SpellLearnSpellNode dbc_node;
-                dbc_node.spell       = spellEffect->EffectTriggerSpell;
-                dbc_node.active      = true;                // all dbc based learned spells is active (show in spell book or hide by client itself)
-
-                // ignore learning nonexistent spells (broken/outdated/or generic learning spell 483
-                if (!sSpellStore.LookupEntry(dbc_node.spell))
-                {
-                    continue;
-                }
-
-                // talent or passive spells or skill-step spells auto-casted and not need dependent learning,
-                // pet teaching spells don't must be dependent learning (casted)
-                // other required explicit dependent learning
-                dbc_node.autoLearned = spellEffect->EffectImplicitTargetA==TARGET_PET || GetTalentSpellCost(spell) > 0 || IsPassiveSpell(entry) || IsSpellHaveEffect(entry,SPELL_EFFECT_SKILL_STEP);
-
-                SpellLearnSpellMapBounds db_node_bounds = GetSpellLearnSpellMapBounds(spell);
-
-                bool found = false;
-                for (SpellLearnSpellMap::const_iterator itr = db_node_bounds.first; itr != db_node_bounds.second; ++itr)
-                {
-                    if (itr->second.spell == dbc_node.spell)
-                    {
-                        sLog.outErrorDb("Spell %u auto-learn spell %u in spell.dbc then the record in `spell_learn_spell` is redundant, please fix DB.",
-                                        spell, dbc_node.spell);
-                        found = true;
-                        break;
-                    }
-                }
-
-                if (!found)                                 // add new spell-spell pair if not found
-                {
-                    mSpellLearnSpells.insert(SpellLearnSpellMap::value_type(spell, dbc_node));
-                    ++dbc_count;
-                }
-            }
-        }
-    }
-
-    sLog.outString(">> Loaded %u spell learn spells + %u found in DBC", count, dbc_count);
-    sLog.outString();
-}
-
-/**
- * @brief Loads and validates scripted spell targets.
- */
-void SpellMgr::LoadSpellScriptTarget()
-{
-    sSpellScriptTargetStorage.Load();
-
-    // Check content
-    for (SQLMultiStorage::SQLSIterator<SpellTargetEntry> itr = sSpellScriptTargetStorage.getDataBegin<SpellTargetEntry>(); itr < sSpellScriptTargetStorage.getDataEnd<SpellTargetEntry>(); ++itr)
-    {
-        SpellEntry const* spellProto = sSpellStore.LookupEntry(itr->spellId);
-        if (!spellProto)
-        {
-            sLog.outErrorDb("Table `spell_script_target`: spellId %u listed for TargetEntry %u does not exist.", itr->spellId, itr->targetEntry);
-            sSpellScriptTargetStorage.EraseEntry(itr->spellId);
-            continue;
-        }
-
-        bool targetfound = false;
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            SpellEffectEntry const* spellEffect = spellProto->GetSpellEffect(SpellEffectIndex(i));
-            if (!spellEffect)
-            {
-                continue;
-            }
-
-            if ( spellEffect->EffectImplicitTargetA == TARGET_SCRIPT ||
-                spellEffect->EffectImplicitTargetB == TARGET_SCRIPT ||
-                spellEffect->EffectImplicitTargetA == TARGET_SCRIPT_COORDINATES ||
-                spellEffect->EffectImplicitTargetB == TARGET_SCRIPT_COORDINATES ||
-                spellEffect->EffectImplicitTargetA == TARGET_FOCUS_OR_SCRIPTED_GAMEOBJECT ||
-                spellEffect->EffectImplicitTargetB == TARGET_FOCUS_OR_SCRIPTED_GAMEOBJECT ||
-                spellEffect->EffectImplicitTargetA == TARGET_AREAEFFECT_INSTANT ||
-                spellEffect->EffectImplicitTargetB == TARGET_AREAEFFECT_INSTANT ||
-                spellEffect->EffectImplicitTargetA == TARGET_AREAEFFECT_CUSTOM ||
-                spellEffect->EffectImplicitTargetB == TARGET_AREAEFFECT_CUSTOM ||
-                spellEffect->EffectImplicitTargetA == TARGET_AREAEFFECT_GO_AROUND_SOURCE ||
-                spellEffect->EffectImplicitTargetB == TARGET_AREAEFFECT_GO_AROUND_SOURCE ||
-                spellEffect->EffectImplicitTargetA == TARGET_AREAEFFECT_GO_AROUND_DEST ||
-                spellEffect->EffectImplicitTargetB == TARGET_AREAEFFECT_GO_AROUND_DEST ||
-                spellEffect->EffectImplicitTargetA == TARGET_NARROW_FRONTAL_CONE ||
-                spellEffect->EffectImplicitTargetB == TARGET_NARROW_FRONTAL_CONE ||
-                spellEffect->EffectImplicitTargetA == TARGET_NARROW_FRONTAL_CONE_2 ||
-                spellEffect->EffectImplicitTargetB == TARGET_NARROW_FRONTAL_CONE_2)
-            {
-                targetfound = true;
-                break;
-            }
-        }
-        if (!targetfound)
-        {
-            sLog.outErrorDb("Table `spell_script_target`: spellId %u listed for TargetEntry %u does not have any implicit target TARGET_SCRIPT(38) or TARGET_SCRIPT_COORDINATES (46) or TARGET_FOCUS_OR_SCRIPTED_GAMEOBJECT (40).", itr->spellId, itr->targetEntry);
-            sSpellScriptTargetStorage.EraseEntry(itr->spellId);
-            continue;
-        }
-
-        if (itr->type >= MAX_SPELL_TARGET_TYPE)
-        {
-            sLog.outErrorDb("Table `spell_script_target`: target type %u for TargetEntry %u is incorrect.", itr->type, itr->targetEntry);
-            sSpellScriptTargetStorage.EraseEntry(itr->spellId);
-            continue;
-        }
-
-        // Checks by target type
-        switch (itr->type)
-        {
-            case SPELL_TARGET_TYPE_GAMEOBJECT:
-            {
-                if (!itr->targetEntry)
-                {
-                    break;
-                }
-
-                if (!sGOStorage.LookupEntry<GameObjectInfo>(itr->targetEntry))
-                {
-                    sLog.outErrorDb("Table `spell_script_target`: gameobject template entry %u does not exist.", itr->targetEntry);
-                    sSpellScriptTargetStorage.EraseEntry(itr->spellId);
-                    continue;
-                }
-                break;
-            }
-            default:
-                if (!itr->targetEntry)
-                {
-                    sLog.outErrorDb("Table `spell_script_target`: target entry == 0 for not GO target type (%u).", itr->type);
-                    sSpellScriptTargetStorage.EraseEntry(itr->spellId);
-                    continue;
-                }
-                if (const CreatureInfo* cInfo = sCreatureStorage.LookupEntry<CreatureInfo>(itr->targetEntry))
-                {
-                    if (itr->spellId == 30427 && !cInfo->SkinningLootId)
-                    {
-                        sLog.outErrorDb("Table `spell_script_target` has creature %u as a target of spellid 30427, but this creature has no SkinningLootId. Gas extraction will not work!", cInfo->Entry);
-                        sSpellScriptTargetStorage.EraseEntry(itr->spellId);
-                        continue;
-                    }
-                }
-                else
-                {
-                    sLog.outErrorDb("Table `spell_script_target`: creature template entry %u does not exist.", itr->targetEntry);
-                    sSpellScriptTargetStorage.EraseEntry(itr->spellId);
-                    continue;
-                }
-                break;
-        }
-    }
-
-    // Check all spells
-    if (!sLog.HasLogFilter(LOG_FILTER_DB_STRICTED_CHECK))
-    {
-        for (uint32 i = 1; i < sSpellStore.GetNumRows(); ++i)
-        {
-            SpellEntry const* spellInfo = sSpellStore.LookupEntry(i);
-            if (!spellInfo)
-            {
-                continue;
-            }
-
-            for (int j = 0; j < MAX_EFFECT_INDEX; ++j)
-            {
-                SpellEffectEntry const* spellEffect = spellInfo->GetSpellEffect(SpellEffectIndex(j));
-                if (!spellEffect)
-                {
-                    continue;
-                }
-
-                if (spellEffect->EffectImplicitTargetA == TARGET_SCRIPT ||
-                    spellEffect->EffectImplicitTargetA != TARGET_SELF && spellEffect->EffectImplicitTargetB == TARGET_SCRIPT)
-                {
-                    SQLMultiStorage::SQLMSIteratorBounds<SpellTargetEntry> bounds = sSpellScriptTargetStorage.getBounds<SpellTargetEntry>(i);
-                    if (bounds.first == bounds.second)
-                    {
-                        sLog.outErrorDb("Spell (ID: %u) has effect EffectImplicitTargetA/EffectImplicitTargetB = %u (TARGET_SCRIPT), but does not have record in `spell_script_target`", spellInfo->Id, TARGET_SCRIPT);
-                        break;                              // effects of spell
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * @brief Loads spell pet aura mappings from the database.
- */
-void SpellMgr::LoadSpellPetAuras()
-{
-    mSpellPetAuraMap.clear();                               // need for reload case
-
-    uint32 count = 0;
-
-    //                                                 0        1           2      3
-    QueryResult* result = WorldDatabase.Query("SELECT `spell`, `effectId`, `pet`, `aura` FROM `spell_pet_auras`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u spell pet auras", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 spell = fields[0].GetUInt32();
-        SpellEffectIndex eff = SpellEffectIndex(fields[1].GetUInt32());
-        uint32 pet = fields[2].GetUInt32();
-        uint32 aura = fields[3].GetUInt32();
-
-        if (eff >= MAX_EFFECT_INDEX)
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_pet_auras` with wrong spell effect index (%u)", spell, eff);
-            continue;
-        }
-
-        SpellPetAuraMap::iterator itr = mSpellPetAuraMap.find((spell << 8) + eff);
-        if (itr != mSpellPetAuraMap.end())
-        {
-            itr->second.AddAura(pet, aura);
-        }
-        else
-        {
-            SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell);
-            if (!spellInfo)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_pet_auras` does not exist", spell);
-                continue;
-            }
-
-            SpellEffectEntry const* spellEffect = spellInfo->GetSpellEffect(eff);
-            if (!spellEffect || spellEffect->Effect != SPELL_EFFECT_DUMMY &&
-               (spellEffect->Effect != SPELL_EFFECT_APPLY_AURA ||
-                spellEffect->EffectApplyAuraName != SPELL_AURA_DUMMY))
-            {
-                sLog.outError("Spell %u listed in `spell_pet_auras` does not have dummy aura or dummy effect", spell);
-                continue;
-            }
-
-            SpellEntry const* spellInfo2 = sSpellStore.LookupEntry(aura);
-            if (!spellInfo2)
-            {
-                sLog.outErrorDb("Aura %u listed in `spell_pet_auras` does not exist", aura);
-                continue;
-            }
-
-            PetAura pa(pet, aura, spellEffect->EffectImplicitTargetA == TARGET_PET, spellEffect->CalculateSimpleValue());
-            mSpellPetAuraMap[(spell << 8) + eff] = pa;
-        }
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u spell pet auras", count);
-    sLog.outString();
-}
-
-void SpellMgr::LoadPetLevelupSpellMap()
-{
-    uint32 count = 0;
-    uint32 family_count = 0;
-
-    for (uint32 i = 0; i < sCreatureFamilyStore.GetNumRows(); ++i)
-    {
-        CreatureFamilyEntry const* creatureFamily = sCreatureFamilyStore.LookupEntry(i);
-        if (!creatureFamily)                                // not exist
-        {
-            continue;
-        }
-
-        for (uint32 j = 0; j < sSkillLineAbilityStore.GetNumRows(); ++j)
-        {
-            SkillLineAbilityEntry const* skillLine = sSkillLineAbilityStore.LookupEntry(j);
-            if (!skillLine)
-            {
-                continue;
-            }
-
-            if (skillLine->skillId != creatureFamily->skillLine[0] &&
-                    (!creatureFamily->skillLine[1] || skillLine->skillId != creatureFamily->skillLine[1]))
-                continue;
-
-            if (skillLine->learnOnGetSkill != ABILITY_LEARNED_ON_GET_RACE_OR_CLASS_SKILL)
-            {
-                continue;
-            }
-
-            SpellEntry const* spell = sSpellStore.LookupEntry(skillLine->spellId);
-            if (!spell)                                     // not exist
-            {
-                continue;
-            }
-
-            PetLevelupSpellSet& spellSet = mPetLevelupSpellMap[creatureFamily->ID];
-            if (spellSet.empty())
-            {
-                ++family_count;
-            }
-
-            spellSet.insert(PetLevelupSpellSet::value_type(spell->GetSpellLevel(),spell->Id));
-            count++;
-        }
-    }
-
-    sLog.outString(">> Loaded %u pet levelup and default spells for %u families", count, family_count);
-    sLog.outString();
-}
-
-bool SpellMgr::LoadPetDefaultSpells_helper(CreatureInfo const* cInfo, PetDefaultSpellsEntry& petDefSpells)
-{
-    // skip empty list;
-    bool have_spell = false;
-    for (int j = 0; j < MAX_CREATURE_SPELL_DATA_SLOT; ++j)
-    {
-        if (petDefSpells.spellid[j])
-        {
-            have_spell = true;
-            break;
-        }
-    }
-    if (!have_spell)
-    {
-        return false;
-    }
-
-    // remove duplicates with levelupSpells if any
-    if (PetLevelupSpellSet const* levelupSpells = cInfo->Family ? GetPetLevelupSpellList(cInfo->Family) : NULL)
-    {
-        for (int j = 0; j < MAX_CREATURE_SPELL_DATA_SLOT; ++j)
-        {
-            if (!petDefSpells.spellid[j])
-            {
-                continue;
-            }
-
-            for (PetLevelupSpellSet::const_iterator itr = levelupSpells->begin(); itr != levelupSpells->end(); ++itr)
-            {
-                if (itr->second == petDefSpells.spellid[j])
-                {
-                    petDefSpells.spellid[j] = 0;
-                    break;
-                }
-            }
-        }
-    }
-
-    // skip empty list;
-    have_spell = false;
-    for (int j = 0; j < MAX_CREATURE_SPELL_DATA_SLOT; ++j)
-    {
-        if (petDefSpells.spellid[j])
-        {
-            have_spell = true;
-            break;
-        }
-    }
-
-    return have_spell;
-}
-
-void SpellMgr::LoadPetDefaultSpells()
-{
-    MANGOS_ASSERT(MAX_CREATURE_SPELL_DATA_SLOT <= CREATURE_MAX_SPELLS);
-
-    mPetDefaultSpellsMap.clear();
-
-    uint32 countCreature = 0;
-    uint32 countData = 0;
-
-    for (uint32 i = 0; i < sCreatureStorage.GetMaxEntry(); ++i)
-    {
-        CreatureInfo const* cInfo = sCreatureStorage.LookupEntry<CreatureInfo>(i);
-        if (!cInfo)
-        {
-            continue;
-        }
-
-        if (!cInfo->PetSpellDataId)
-        {
-            continue;
-        }
-
-        // for creature with PetSpellDataId get default pet spells from dbc
-        CreatureSpellDataEntry const* spellDataEntry = sCreatureSpellDataStore.LookupEntry(cInfo->PetSpellDataId);
-        if (!spellDataEntry)
-        {
-            continue;
-        }
-
-        int32 petSpellsId = -(int32)cInfo->PetSpellDataId;
-        PetDefaultSpellsEntry petDefSpells;
-        for (int j = 0; j < MAX_CREATURE_SPELL_DATA_SLOT; ++j)
-        {
-            petDefSpells.spellid[j] = spellDataEntry->spellId[j];
-        }
-
-        if (LoadPetDefaultSpells_helper(cInfo, petDefSpells))
-        {
-            mPetDefaultSpellsMap[petSpellsId] = petDefSpells;
-            ++countData;
-        }
-    }
-
-    // different summon spells
-    for (uint32 i = 0; i < sSpellStore.GetNumRows(); ++i)
-    {
-        SpellEntry const* spellEntry = sSpellStore.LookupEntry(i);
-        if (!spellEntry)
-        {
-            continue;
-        }
-
-        for (int k = 0; k < MAX_EFFECT_INDEX; ++k)
-        {
-            SpellEffectEntry const* spellEffect = spellEntry->GetSpellEffect(SpellEffectIndex(k));
-            if (!spellEffect)
-            {
-                continue;
-            }
-
-            if (spellEffect->Effect==SPELL_EFFECT_SUMMON || spellEffect->Effect==SPELL_EFFECT_SUMMON_PET)
-            {
-                uint32 creature_id = spellEffect->EffectMiscValue;
-                CreatureInfo const* cInfo = sCreatureStorage.LookupEntry<CreatureInfo>(creature_id);
-                if (!cInfo)
-                {
-                    continue;
-                }
-
-                // already loaded
-                if (cInfo->PetSpellDataId)
-                {
-                    continue;
-                }
-
-                // for creature without PetSpellDataId get default pet spells from creature_template
-                int32 petSpellsId = cInfo->Entry;
-                if (mPetDefaultSpellsMap.find(cInfo->Entry) != mPetDefaultSpellsMap.end())
-                {
-                    continue;
-                }
-
-                PetDefaultSpellsEntry petDefSpells;
-                if (CreatureTemplateSpells const* templateSpells = sCreatureTemplateSpellsStorage.LookupEntry<CreatureTemplateSpells>(cInfo->Entry))
-                    for (int j = 0; j < MAX_CREATURE_SPELL_DATA_SLOT; ++j)
-                    {
-                        petDefSpells.spellid[j] = templateSpells->spells[j];
-                    }
-
-                if (LoadPetDefaultSpells_helper(cInfo, petDefSpells))
-                {
-                    mPetDefaultSpellsMap[petSpellsId] = petDefSpells;
-                    ++countCreature;
-                }
-            }
-        }
-    }
-
-    sLog.outString(">> Loaded addition spells for %u pet spell data entries and %u summonable creature templates", countData, countCreature);
-    sLog.outString();
-}
-
 /// Some checks for spells, to prevent adding deprecated/broken spells for trainers, spell book, etc
 bool SpellMgr::IsSpellValid(SpellEntry const* spellInfo, Player* pl, bool msg)
 {
@@ -4969,239 +3049,6 @@ bool SpellMgr::IsSpellValid(SpellEntry const* spellInfo, Player* pl, bool msg)
     }
 
     return true;
-}
-
-/**
- * @brief Loads spell area requirement records from the database.
- */
-void SpellMgr::LoadSpellAreas()
-{
-    mSpellAreaMap.clear();                                  // need for reload case
-    mSpellAreaForAuraMap.clear();
-
-    uint32 count = 0;
-
-    //                                                0      1     2            3                   4          5             6           7         8       9
-    QueryResult* result = WorldDatabase.Query("SELECT `spell`, `area`, `quest_start`, `quest_start_active`, `quest_end`, `condition_id`, `aura_spell`, `racemask`, `gender`, `autocast` FROM `spell_area`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u spell area requirements", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 spell = fields[0].GetUInt32();
-        SpellArea spellArea;
-        spellArea.spellId             = spell;
-        spellArea.areaId              = fields[1].GetUInt32();
-        spellArea.questStart          = fields[2].GetUInt32();
-        spellArea.questStartCanActive = fields[3].GetBool();
-        spellArea.questEnd            = fields[4].GetUInt32();
-        spellArea.conditionId         = fields[5].GetUInt16();
-        spellArea.auraSpell           = fields[6].GetInt32();
-        spellArea.raceMask            = fields[7].GetUInt32();
-        spellArea.gender              = Gender(fields[8].GetUInt8());
-        spellArea.autocast            = fields[9].GetBool();
-
-        if (!sSpellStore.LookupEntry(spell))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_area` does not exist", spell);
-            continue;
-        }
-
-        {
-            bool ok = true;
-            SpellAreaMapBounds sa_bounds = GetSpellAreaMapBounds(spellArea.spellId);
-            for (SpellAreaMap::const_iterator itr = sa_bounds.first; itr != sa_bounds.second; ++itr)
-            {
-                if (spellArea.spellId != itr->second.spellId)
-                {
-                    continue;
-                }
-                if (spellArea.areaId != itr->second.areaId)
-                {
-                    continue;
-                }
-                if (spellArea.questStart != itr->second.questStart)
-                {
-                    continue;
-                }
-                if (spellArea.auraSpell != itr->second.auraSpell)
-                {
-                    continue;
-                }
-                if ((spellArea.raceMask & itr->second.raceMask) == 0)
-                {
-                    continue;
-                }
-                if (spellArea.gender != itr->second.gender)
-                {
-                    continue;
-                }
-
-                // duplicate by requirements
-                ok = false;
-                break;
-            }
-
-            if (!ok)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_area` already listed with similar requirements.", spell);
-                continue;
-            }
-        }
-
-        if (spellArea.areaId && !GetAreaEntryByAreaID(spellArea.areaId))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_area` have wrong area (%u) requirement", spell, spellArea.areaId);
-            continue;
-        }
-
-        if (spellArea.conditionId && !sConditionStorage.LookupEntry<PlayerCondition>(spellArea.conditionId))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_area` have wrong conditionId (%u) requirement", spell, spellArea.conditionId);
-            continue;
-        }
-        else if (!spellArea.conditionId)
-        {
-            if (spellArea.questStart && !sObjectMgr.GetQuestTemplate(spellArea.questStart))
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_area` have wrong start quest (%u) requirement", spell, spellArea.questStart);
-                continue;
-            }
-
-            if (spellArea.questEnd)
-            {
-                if (!sObjectMgr.GetQuestTemplate(spellArea.questEnd))
-                {
-                    sLog.outErrorDb("Spell %u listed in `spell_area` have wrong end quest (%u) requirement", spell, spellArea.questEnd);
-                    continue;
-                }
-
-                if (spellArea.questEnd == spellArea.questStart && !spellArea.questStartCanActive)
-                {
-                    sLog.outErrorDb("Spell %u listed in `spell_area` have quest (%u) requirement for start and end in same time", spell, spellArea.questEnd);
-                    continue;
-                }
-            }
-
-            if (spellArea.raceMask && (spellArea.raceMask & RACEMASK_ALL_PLAYABLE) == 0)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_area` have wrong race mask (%u) requirement", spell, spellArea.raceMask);
-                continue;
-            }
-
-            if (spellArea.gender != GENDER_NONE && spellArea.gender != GENDER_FEMALE && spellArea.gender != GENDER_MALE)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_area` have wrong gender (%u) requirement", spell, spellArea.gender);
-                continue;
-            }
-        }
-
-        if (spellArea.auraSpell)
-        {
-            SpellEntry const* spellInfo = sSpellStore.LookupEntry(abs(spellArea.auraSpell));
-            if (!spellInfo)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_area` have wrong aura spell (%u) requirement", spell, abs(spellArea.auraSpell));
-                continue;
-            }
-
-            SpellEffectEntry const* spellEffect = spellInfo->GetSpellEffect(EFFECT_INDEX_0);
-            if (!spellEffect)
-            {
-                continue;
-            }
-
-            switch (spellEffect->EffectApplyAuraName)
-            {
-                case SPELL_AURA_DUMMY:
-                case SPELL_AURA_PHASE:
-                case SPELL_AURA_PHASE_2:
-                case SPELL_AURA_GHOST:
-                    break;
-                default:
-                    sLog.outErrorDb("Spell %u listed in `spell_area` have aura spell requirement (%u) without dummy/phase/ghost aura in effect 0", spell, abs(spellArea.auraSpell));
-                    continue;
-            }
-
-            if (uint32(abs(spellArea.auraSpell)) == spellArea.spellId)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_area` have aura spell (%u) requirement for itself", spell, abs(spellArea.auraSpell));
-                continue;
-            }
-
-            // not allow autocast chains by auraSpell field (but allow use as alternative if not present)
-            if (spellArea.autocast && spellArea.auraSpell > 0)
-            {
-                bool chain = false;
-                SpellAreaForAuraMapBounds saBound = GetSpellAreaForAuraMapBounds(spellArea.spellId);
-                for (SpellAreaForAuraMap::const_iterator itr = saBound.first; itr != saBound.second; ++itr)
-                {
-                    if (itr->second->autocast && itr->second->auraSpell > 0)
-                    {
-                        chain = true;
-                        break;
-                    }
-                }
-
-                if (chain)
-                {
-                    sLog.outErrorDb("Spell %u listed in `spell_area` have aura spell (%u) requirement that itself autocast from aura", spell, spellArea.auraSpell);
-                    continue;
-                }
-
-                SpellAreaMapBounds saBound2 = GetSpellAreaMapBounds(spellArea.auraSpell);
-                for (SpellAreaMap::const_iterator itr2 = saBound2.first; itr2 != saBound2.second; ++itr2)
-                {
-                    if (itr2->second.autocast && itr2->second.auraSpell > 0)
-                    {
-                        chain = true;
-                        break;
-                    }
-                }
-
-                if (chain)
-                {
-                    sLog.outErrorDb("Spell %u listed in `spell_area` have aura spell (%u) requirement that itself autocast from aura", spell, spellArea.auraSpell);
-                    continue;
-                }
-            }
-        }
-
-        SpellArea const* sa = &mSpellAreaMap.insert(SpellAreaMap::value_type(spell, spellArea))->second;
-
-        // for search by current zone/subzone at zone/subzone change
-        if (spellArea.areaId)
-        {
-            mSpellAreaForAreaMap.insert(SpellAreaForAreaMap::value_type(spellArea.areaId, sa));
-        }
-
-        // for search at aura apply
-        if (spellArea.auraSpell)
-        {
-            mSpellAreaForAuraMap.insert(SpellAreaForAuraMap::value_type(abs(spellArea.auraSpell), sa));
-        }
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u spell area requirements", count);
-    sLog.outString();
 }
 
 /**
@@ -5380,67 +3227,6 @@ SpellCastResult SpellMgr::GetSpellAllowedInLocationError(SpellEntry const* spell
     }
 
     return SPELL_CAST_OK;
-}
-
-/**
- * @brief Builds the skill-line ability multimap from DBC data.
- */
-void SpellMgr::LoadSkillLineAbilityMap()
-{
-    mSkillLineAbilityMap.clear();
-
-    BarGoLink bar(sSkillLineAbilityStore.GetNumRows());
-    uint32 count = 0;
-
-    for (uint32 i = 0; i < sSkillLineAbilityStore.GetNumRows(); ++i)
-    {
-        bar.step();
-        SkillLineAbilityEntry const* SkillInfo = sSkillLineAbilityStore.LookupEntry(i);
-        if (!SkillInfo)
-        {
-            continue;
-        }
-
-        mSkillLineAbilityMap.insert(SkillLineAbilityMap::value_type(SkillInfo->spellId, SkillInfo));
-        ++count;
-    }
-
-    sLog.outString(">> Loaded %u SkillLineAbility MultiMap Data", count);
-    sLog.outString();
-}
-
-/**
- * @brief Builds the skill race/class requirement multimap from DBC data.
- */
-void SpellMgr::LoadSkillRaceClassInfoMap()
-{
-    mSkillRaceClassInfoMap.clear();
-
-    BarGoLink bar(sSkillRaceClassInfoStore.GetNumRows());
-    uint32 count = 0;
-
-    for (uint32 i = 0; i < sSkillRaceClassInfoStore.GetNumRows(); ++i)
-    {
-        bar.step();
-        SkillRaceClassInfoEntry const* skillRCInfo = sSkillRaceClassInfoStore.LookupEntry(i);
-        if (!skillRCInfo)
-        {
-            continue;
-        }
-
-        // not all skills really listed in ability skills list
-        if (!sSkillLineStore.LookupEntry(skillRCInfo->skillId))
-        {
-            continue;
-        }
-
-        mSkillRaceClassInfoMap.insert(SkillRaceClassInfoMap::value_type(skillRCInfo->skillId, skillRCInfo));
-
-        ++count;
-    }
-
-    sLog.outString(">> Loaded %u SkillRaceClassInfo MultiMap Data", count);
-    sLog.outString();
 }
 
 /**
